refactor: split FormPreviewSection component
