+++ src/features/form/components/FormDetail/components/FormPreviewSection/FormPreviewSection.module.css
@@ -1,0 +1,66 @@
+.previewSection {
+	display: flex;
+	flex-direction: column;
+	gap: 4rem;
+}
+
+.previewBlock {
+	display: flex;
+	flex-direction: column;
+	gap: 1rem;
+}
+
+.previewHeader {
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+	padding-bottom: 0.75rem;
+}
+
+.previewTitle {
+	font-size: 1.6rem;
+	font-weight: 520;
+	color: var(--color-heading);
+	margin: 0;
+}
+
+.previewList {
+	list-style: disc;
+	padding-left: 1.5rem;
+	margin: 0;
+	display: flex;
+	flex-direction: column;
+	gap: 0.75rem;
+}
+
+.previewList li {
+	font-size: 1rem;
+	color: var(--color-body);
+	line-height: 1.5;
+}
+
+.previewList li strong {
+	color: var(--color-heading);
+	font-weight: 600;
+	margin-right: 0.5rem;
+}
+
+.previewSectionTitle {
+	margin: 0;
+}
+
+.previewAnswerLabel {
+	font-weight: 500;
+}
+
+.previewAnswerEmpty {
+	color: var(--red);
+}
+
+.requiredAsterisk {
+	color: var(--red);
+}
+
+.caption {
+	color: var(--color-caption);
+}
